fixed more -Wshorten-64-to-32 warnings

+++ src/pdft_func/XCIntegratorForPDFT.cpp
@@ -210,7 +210,7 @@ integrateVxcPDFTForLDA(CAOKohnShamMatrix&              aoFockMatrix,
 
             // go through GTO blocks
 
-            for (size_t i_block = 0, idx = 0; i_block < n_gto_blocks; i_block++)
+            for (int i_block = 0, idx = 0; i_block < static_cast<int>(n_gto_blocks); i_block++)
             {
                 const auto& gto_block = ptr_gto_blocks[i_block];
 
@@ -475,7 +475,7 @@ integrateVxcPDFTForGGA(CAOKohnShamMatrix&              aoFockMatrix,
 
             // go through GTO blocks
 
-            for (size_t i_block = 0, idx = 0; i_block < n_gto_blocks; i_block++)
+            for (int i_block = 0, idx = 0; i_block < static_cast<int>(n_gto_blocks); i_block++)
             {
                 const auto& gto_block = ptr_gto_blocks[i_block];
